Support Wayland and improve Docker compatibility
Added Wayland support in the Linux GLFW integration, enabling compatibility with both X11 and Wayland environments. Enhanced the Docker configuration with Wayland-specific libraries, Weston setup, and related scripts for better testing flexibility. Minor fixes and optimizations in build and run scripts were also applied.

+++ docker/build.sh
@@ -1,4 +1,7 @@
 CHROME_URL="https://dl.google.com/linux/deb/pool/main/g/google-chrome-stable/google-chrome-stable_128.0.6613.84-1_amd64.deb"
 SCRIPT_DIR=$(dirname "$(realpath "$0")")
-curl -o $SCRIPT_DIR/chrome-128.deb $CHROME_URL
-docker build -t ubuntu-wgpu4k $SCRIPT_DIR
+# Vérifier si le fichier Chrome existe déjà
+if [ ! -f $SCRIPT_DIR/chrome-128.deb ]; then
+  curl -o $SCRIPT_DIR/chrome-128.deb $CHROME_URL
+fi
+docker build --platform=linux/amd64 -t ubuntu-wgpu4k $SCRIPT_DIR
